// src/app/shared/interfaces/terminal.interface.ts
export interface TerminalResponse {
    id?:           string;
    nombre?:       string;
    descripcion?:  string;
    precio?:       number;
    comDebito?:    number;
    comCredito?:   number;
    comInternac?:  number;
    rentaMes?:     number;
    cadena64?:     string;
    tipoProducto?: null;
}